requirements.txt
kafka-python
psycopg2-binary
faker

fastapi==0.128.0
uvicorn[standard]==0.40.0
asyncpg==0.31.0
Jinja2==3.1.6
python-dotenv==1.2.1